Move setDoneCheckboxes function from displayTodos.js into todosController.js
