+++ src/javascript/startApp.js
@@ -5,7 +5,7 @@ import { setToggleMenus } from './toggleMenus.js';
 import { setToggleForms } from './toggleForms.js';
 import setShowAllCategoriesBtn from './showAllCategories.js';
 import setCategoriesController from './categoriesController.js';
-import setTodosController from './todosController.js';
+import { setTodosController } from './todosController.js';
 import { displayAllTodos } from './displayTodos.js';
 
 const start = () => {
