+++ src/javascript/todosController.js
@@ -1,4 +1,4 @@
-import { displayTodo } from './displayTodos.js';
+import { displayTodo } from './displayTodos.js'; // eslint-disable-line
 import {
   addTodoToArray, categoriesArray, deleteTodoFromArray, editTodoInArray,
 } from './database.js';
@@ -86,6 +86,27 @@ const setDeleteTodoListener = () => {
   });
 };
 
+const setDoneCheckboxes = () => {
+  let doneCheckboxes = document.querySelectorAll('.done-checkbox');
+
+  doneCheckboxes.forEach((doneCheckbox) => {
+    doneCheckbox.replaceWith(doneCheckbox.cloneNode(true));
+  });
+
+  doneCheckboxes = document.querySelectorAll('.done-checkbox');
+
+  doneCheckboxes.forEach(((doneCheckbox) => {
+    doneCheckbox.addEventListener('change', (e) => {
+      const currentTodoHeading = e.currentTarget.parentElement.parentElement
+        .parentElement.childNodes[0];
+      const currentTodoDescription = e.currentTarget.parentElement.parentElement
+        .parentElement.childNodes[1];
+      currentTodoHeading.classList.toggle('done');
+      currentTodoDescription.classList.toggle('done');
+    });
+  }));
+};
+
 const deleteAllDoneTodos = () => {
 
 };
@@ -96,4 +117,4 @@ const setTodosController = () => {
   setDeleteTodoListener();
 };
 
-export default setTodosController;
+export { setTodosController, setDoneCheckboxes };
